bot/requirements.txt
python-telegram-bot
pandas_datareader
tqdm
polygon
pandas_ta
pandas
